requirements.txt
discord.py==2.1.0
packaging==22.0
pydantic==1.10.2
tomlkit==0.11.6